+++ crypto/lib/crypto.dart
@@ -1,0 +1,3 @@
+library;
+
+export 'src/turnkey.dart' show decryptCredentialBundle, decryptExportBundle, encryptPrivateKeyToBundle, encryptWalletToBundle;
