+++ crypto/lib/src/constant.dart
