+++ crypto/pubspec.yaml
@@ -1,0 +1,22 @@
+name: turnkey_crypto
+description: This package consolidates some common cryptographic utilities used across our applications, particularly primitives related to keys, encryption, and decryption in a pure Dart implementation.
+version: 1.0.0
+repository: https://github.com/tkhq/dart-sdk
+
+environment:
+  sdk: 3.6.1
+
+dependencies:
+  turnkey_encoding:
+    path: ../encoding # Encoding package needs to be published, so for now we just use a local path
+  pointycastle: 3.9.0
+  base58check: 2.0.0
+  bs58: 1.0.2
+  cryptography: 2.7.0
+
+
+dev_dependencies:
+  lints: 5.0.0
+  test: 1.25.0
+
+publish_to: none # We add this temporarily, so we can use local dependencies
