--- encoding/analysis_options.yaml
@@ -1,1 +1,0 @@
-include: package:lints/recommended.yaml
