+++ encoding/lib/src/encoding.dart
