+++ encoding/pubspec.yaml
@@ -1,10 +1,12 @@
-name: encoding
-description: A starting point for Dart libraries or applications.
+name: turnkey_encoding
+description: This is a package containing decoding and encoding functions. This is used by other Turnkey packages. Feel free to import exported functions into your own apps, but be aware that there may be significant interface changes in the future.
+repository: https://github.com/tkhq/dart-sdk
+
 version: 1.0.0
 
 environment:
-  sdk: ^3.6.1
+  sdk: 3.6.1
 
 dev_dependencies:
-  lints: ^5.0.0
-  test: ^1.24.0
+  lints: 5.0.0
+  test: 1.25.0
