--- turnkey-crypto/lib/crypto.dart
@@ -1,3 +1,0 @@
-library;
-
-export 'src/math.dart' show modSqrt;
