--- turnkey-crypto/pubspec.yaml
@@ -1,23 +1,0 @@
-name: turnkey_crypto
-description: A starting point for Dart libraries or applications.
-version: 1.0.0
-# repository: https://github.com/my_org/my_repo
-
-environment:
-  sdk: ^3.6.1
-
-# Add regular dependencies here.
-dependencies:
-  encoding:
-    path: ../encoding
-  pointycastle: ^3.9.0
-  base58check: ^2.0.0
-  bs58: ^1.0.2
-  cryptography: ^2.7.0
-
-
-dev_dependencies:
-  lints: ^5.0.0
-  test: ^1.24.0
-
-publish_to: none # We add this temporarlly, so we can use local dependencies
